requirements.txt
streamlit
email-validator
pillow
geopy
requests